+++ src/main/java/com/evacipated/cardcrawl/modthespire/patcher/InsertPatchInfo.java
@@ -73,6 +73,11 @@ public class InsertPatchInfo extends PatchInfo
             && returnType.equals(returnType.getClassPool().get(SpireReturn.class.getName()))) {
 
             hasEarlyReturn = true;
+        } else if (ctMethodToPatch instanceof CtConstructor
+            && !returnType.equals(CtPrimitiveType.voidType)
+            && returnType.equals(returnType.getClassPool().get(SpireReturn.class.getName()))) {
+
+            hasEarlyReturn = true;
         }
 
         CtClass[] insertParamTypes = patchMethod.getParameterTypes();
@@ -152,7 +157,7 @@ public class InsertPatchInfo extends PatchInfo
 
         if (hasEarlyReturn) {
             String earlyReturn = "if (opt.isPresent()) { return";
-            if (!((CtMethod) ctMethodToPatch).getReturnType().equals(CtPrimitiveType.voidType)) {
+            if (ctMethodToPatch instanceof CtMethod && !((CtMethod) ctMethodToPatch).getReturnType().equals(CtPrimitiveType.voidType)) {
                 CtClass toPatchReturnType = ((CtMethod) ctMethodToPatch).getReturnType();
                 String toPatchReturnTypeName = toPatchReturnType.getName();
                 if (toPatchReturnType.isPrimitive()) {